+++ src/main/java/co/nvqa/common/core/cucumber/CoreStandardSteps.java
@@ -7,4 +7,8 @@ import co.nvqa.common.cucumber.glue.StandardSteps;
 public abstract class CoreStandardSteps extends StandardSteps<StandardScenarioManager> implements
     CoreScenarioStorageKeys {
 
+  @Override
+  public void init() {
+    // This method is empty by design.
+  }
 }