+++ src/main/java/co/nvqa/common/core/cucumber/glue/ApiRouteSteps.java
@@ -8,6 +8,7 @@ import co.nvqa.common.core.model.coverage.CreateCoverageRequest;
 import co.nvqa.common.core.model.coverage.CreateCoverageResponse;
 import co.nvqa.common.core.model.other.CoreExceptionResponse;
 import co.nvqa.common.core.model.other.CoreExceptionResponse.Error;
+import co.nvqa.common.core.model.pickup.MilkRunGroup;
 import co.nvqa.common.core.model.reservation.BulkRouteReservationResponse;
 import co.nvqa.common.core.model.route.AddParcelToRouteRequest;
 import co.nvqa.common.core.model.route.AddPickupJobToRouteRequest;
@@ -33,12 +34,14 @@ import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.Date;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import javax.inject.Inject;
 import lombok.Getter;
+import org.apache.commons.lang3.StringUtils;
 import org.assertj.core.api.Assertions;
 
 @ScenarioScoped
@@ -48,10 +51,6 @@ public class ApiRouteSteps extends CoreStandardSteps {
   @Getter
   private RouteClient routeClient;
 
-  @Override
-  public void init() {
-
-  }
 
   /**
    * Sample:
@@ -184,7 +183,7 @@ public class ApiRouteSteps extends CoreStandardSteps {
 
   @Given("API Core - Operator remove pickup job id {string} from route")
   public void apiOperatorRemovePickupJobFromRouteUsingDataBelow(String paJobId) {
-    final Long jobId = Long.parseLong(resolveValue(paJobId));
+    final long jobId = Long.parseLong(resolveValue(paJobId));
     doWithRetry(
         () -> getRouteClient().removePAJobFromRoute(jobId),
         "remove pa job from route");
@@ -329,11 +328,11 @@ public class ApiRouteSteps extends CoreStandardSteps {
           reservationId, overwrite);
 
       Assertions.assertThat(r.statusCode())
-          .withFailMessage("unexpected http status: " + r.statusCode())
+          .as("expected http status: " + r.statusCode())
           .isEqualTo(expectedStatusCode);
 
       Assertions.assertThat(r.getBody().asString())
-          .withFailMessage("unexpected error message: " + r.getBody().asString())
+          .as("expected error message: " + r.getBody().asString())
           .isEqualTo(expectedErrorMessage);
     }, "(expected) failed add reservation to route");
   }
@@ -365,11 +364,11 @@ public class ApiRouteSteps extends CoreStandardSteps {
           reservationResultId);
 
       Assertions.assertThat(r.statusCode())
-          .withFailMessage("unexpected http status: " + r.statusCode())
+          .as("expected http status: " + r.statusCode())
           .isEqualTo(expectedStatusCode);
 
       Assertions.assertThat(r.getBody().asString())
-          .withFailMessage("unexpected error message: " + r.getBody().asString())
+          .as("expected error message: " + r.getBody().asString())
           .isEqualTo(expectedErrorMessage);
     }, "(expected) failed add reservation to route");
   }
@@ -463,10 +462,10 @@ public class ApiRouteSteps extends CoreStandardSteps {
     List<Data> actual =
         fromJson(resolvedDataTable.get("actualResponse"), MergeWaypointsResponse.class).getData();
     Assertions.assertThat(actual)
-        .withFailMessage("merge waypoints response is null")
+        .as("merge waypoints response is not null")
         .isNotNull();
     Assertions.assertThat(actual.size())
-        .withFailMessage("merge waypoints response size doesnt match")
+        .as("merge waypoints response size match")
         .isEqualTo(expected.size());
     expected.forEach(o -> DataEntity.assertListContains(actual, o, "merged waypoints list"));
   }
@@ -480,8 +479,8 @@ public class ApiRouteSteps extends CoreStandardSteps {
   @Given("API Core - Operator new add parcel to DP holding route:")
   public void operatorAddToDpHoldingRoute(Map<String, String> data) {
     data = resolveKeyValues(data);
-    final Long routeId = Long.parseLong(data.get("routeId"));
-    final Long orderId = Long.parseLong(data.get("orderId"));
+    final long routeId = Long.parseLong(data.get("routeId"));
+    final long orderId = Long.parseLong(data.get("orderId"));
     doWithRetry(
         () -> getRouteClient().addToRouteDp(orderId, routeId),
         "Add to route dp order");
@@ -705,9 +704,21 @@ public class ApiRouteSteps extends CoreStandardSteps {
         finalData.get("requestBody")), "add references to Route Group", 2000, 3);
   }
 
+  /**
+   * Sample:
+   * <p>
+   * API Route - create route group: |name|ARG-{uniqueString} | |description|This Route Group is
+   * created by automation test from Operator V2.|
+   *
+   * @param data Map of data from feature file.
+   */
   @Given("API Route - create route group:")
   public void apiOperatorCreateNewRouteGroup(Map<String, String> data) {
     RouteGroup routeGroup = new RouteGroup(resolveKeyValues(data));
+    String uniqueString = CoreTestUtils.generateUniqueId();
+    if (StringUtils.endsWithIgnoreCase(routeGroup.getName(), "{uniqueString}")) {
+      routeGroup.setName(routeGroup.getName().replace("{uniqueString}", uniqueString));
+    }
     doWithRetry(() -> {
       var response = getRouteClient().createRouteGroup(routeGroup);
       putInList(KEY_LIST_OF_CREATED_ROUTE_GROUPS, response);
@@ -720,5 +731,20 @@ public class ApiRouteSteps extends CoreStandardSteps {
     routeIds.stream()
         .map(Long::parseLong)
         .forEach(id -> getRouteClient().deleteRoute(id));
+  }
+
+  @And("API Route - Operator get created Reservation Group params:")
+  public void apiOperatorGetCreatedReservationGroupParams(Map<String, String> dataTableAsMap) {
+    Map<String, String> resolvedDataTable = resolveKeyValues(dataTableAsMap);
+    final String reservationGroupName = resolvedDataTable.get("reservationGroupName");
+    doWithRetry(() -> {
+      List<MilkRunGroup> milkrunGroups = getRouteClient().getMilkrunGroups(new Date());
+
+      MilkRunGroup group = milkrunGroups.stream().filter(
+              milkrunGroup -> StringUtils.equals(milkrunGroup.getName(), reservationGroupName))
+          .findFirst().orElseThrow(() -> new RuntimeException(
+              "Could not find milkrun group with name [" + reservationGroupName + "]"));
+      put(KEY_CORE_CREATED_RESERVATION_GROUP_ID, group.getId());
+    }, "Operator get created Reservation Group params");
   }
 }