+++ src/main/java/co/nvqa/common/core/cucumber/glue/DbCoreSteps.java
@@ -99,9 +99,6 @@ public class DbCoreSteps extends CoreStandardSteps {
   @Inject
   private OrderDao orderDao;
 
-  @Override
-  public void init() {
-  }
 
   @And("DB Core - get Reservation data from reservation id {string}")
   public void coreGetReservationDataFromReservationId(String reservationId) {
@@ -277,6 +274,17 @@ public class DbCoreSteps extends CoreStandardSteps {
           OrderDetails orderDetails = orderDetailsDao.getOrderDetailsByOrderId(resolvedOrderId);
           putInList(KEY_CORE_LIST_OF_ORDER_DETAILS, orderDetails);
         }, "get order details");
+  }
+
+  @When("DB Core - operator get transaction records with:")
+  public void getOrderDetailsByOrderId(Map<String, String> data) {
+    Map<String, String> resolvedData = resolveKeyValues(data);
+    doWithRetry(
+        () -> {
+          List<Transactions> result = transactionsDao.findTransactionByOrderIdAndType(
+              Long.parseLong((resolvedData.get("order_id"))), resolvedData.get("type"));
+          put(KEY_CORE_LIST_OF_TRANSACTIONS, result);
+        }, "get transaction records");
   }
 
   @When("DB Core - verify transactions record:")
@@ -558,7 +566,9 @@ public class DbCoreSteps extends CoreStandardSteps {
               String.class);
           List<String> expectedTags = fromJsonToList(f("[%s]", resolvedDataTable.get("tagIds")),
               String.class);
-          Assertions.assertThat(actualTags.containsAll(expectedTags));
+          Assertions.assertThat(actualTags)
+              .as("Actual tags contains the expected tags: {}", expectedTags)
+              .containsAll(expectedTags);
         }, "verify order_tags_search");
   }