+++ src/main/java/co/nvqa/common/core/utils/EventDetailDeserializer.java
@@ -98,41 +98,35 @@ public class EventDetailDeserializer extends JsonDeserializer<EventDetail> {
     if (node.get("mode") != null) {
       result.setMode(node.get("mode").asText());
     }
-    if (node.get("weight") != null) {
-      if (node.get("weight").isObject()) {
+    if (node.get("weight") != null && (node.get("weight").isObject())) {
         result.setWeight(
             JsonUtils.fromJsonSnakeCase(node.get("weight").toString(), EventValue.class));
-      }
+
     }
-    if (node.get("length") != null) {
-      if (node.get("length").isObject()) {
+    if (node.get("length") != null && (node.get("length").isObject())) {
         result.setLength(
             JsonUtils.fromJsonSnakeCase(node.get("length").toString(), EventValue.class));
-      }
+
     }
-    if (node.get("height") != null) {
-      if (node.get("height").isObject()) {
+    if (node.get("height") != null && (node.get("height").isObject())) {
         result.setHeight(
             JsonUtils.fromJsonSnakeCase(node.get("height").toString(), EventValue.class));
-      }
+
     }
-    if (node.get("width") != null) {
-      if (node.get("width").isObject()) {
+    if (node.get("width") != null && (node.get("width").isObject())) {
         result.setWidth(
             JsonUtils.fromJsonSnakeCase(node.get("width").toString(), EventValue.class));
-      }
+
     }
-    if (node.get("granular_status") != null) {
-      if (node.get("granular_status").isObject()) {
+    if (node.get("granular_status") != null && (node.get("granular_status").isObject())) {
         result.setGranularStatus(
             JsonUtils.fromJsonSnakeCase(node.get("granular_status").toString(), EventValueString.class));
-      }
+
     }
-    if (node.get("order_status") != null) {
-      if (node.get("order_status").isObject()) {
+    if (node.get("order_status") != null && (node.get("order_status").isObject())) {
         result.setOrderStatus(
             JsonUtils.fromJsonSnakeCase(node.get("order_status").toString(), EventValueString.class));
-      }
+
     }
     return result;
   }